+++ lib/slack-ruby-bot-server-events/api/endpoints/slack/actions_endpoint.rb
@@ -14,7 +14,7 @@ module SlackRubyBotServer
                 optional :type, type: String
                 optional :trigger_id, type: String
                 optional :response_url, type: String
-                requires :channel, type: Hash do
+                optional :channel, type: Hash do
                   requires :id, type: String
                   optional :name, type: String
                 end
@@ -31,7 +31,7 @@ module SlackRubyBotServer
                 end
                 optional :message, type: Hash do
                   requires :type, type: String
-                  requires :user, type: String
+                  optional :user, type: String
                   requires :ts, type: String
                   requires :text, type: String
                 end
@@ -39,8 +39,9 @@ module SlackRubyBotServer
             end
             post '/action' do
               action = SlackRubyBotServer::Events::Requests::Action.new(params, request)
+              payload_type = params[:payload][:type]
               callback_id = params[:payload][:callback_id]
-              SlackRubyBotServer::Events.config.run_callbacks(:action, callback_id, action) || body(false)
+              SlackRubyBotServer::Events.config.run_callbacks(:action, [payload_type, callback_id].compact, action) || body(false)
             end
           end
         end