+++ spec/slack-ruby-bot-server-events/api/endpoints/slack/actions_endpoint_spec.rb
@@ -19,6 +19,7 @@ describe SlackRubyBotServer::Events::Api::Endpoints::Slack::ActionsEndpoint do
 
     let(:payload) do
       {
+        type: 'message_action',
         actions: [{ name: 'id', value: '43749' }],
         channel: { id: 'C12345', name: 'channel' },
         user: { id: 'user_id' },
@@ -32,77 +33,55 @@ describe SlackRubyBotServer::Events::Api::Endpoints::Slack::ActionsEndpoint do
       before do
         SlackRubyBotServer::Events.configure do |config|
           config.on :action do |action|
-            if action[:payload][:callback_id] == 'action_id'
-              { text: 'Success!' }
-            else
-              raise "Action #{action[:payload][:callback_id]} is not supported."
-            end
+            { text: "#{action[:payload][:type]}/#{action[:payload][:callback_id]}" }
           end
         end
       end
 
       it 'performs action' do
-        post '/api/slack/action', payload: payload.merge(callback_id: 'action_id').to_json
+        post '/api/slack/action', payload: payload.to_json
         expect(last_response.status).to eq 201
         response = JSON.parse(last_response.body)
-        expect(response).to eq('text' => 'Success!')
-      end
-
-      it 'errors on an unhandled action' do
-        post '/api/slack/action', payload: payload.merge(callback_id: 'invalid').to_json
-        expect(last_response.status).to eq 400
-        response = JSON.parse(last_response.body)
-        expect(response['message']).to eq('Action invalid is not supported.')
+        expect(response).to eq('text' => 'message_action/action_id')
       end
     end
 
-    context 'with a specific action handler' do
+    context 'with a message type action handler' do
       before do
         SlackRubyBotServer::Events.configure do |config|
-          config.on :action, 'action_id' do |_action|
-            { text: 'Success!' }
+          config.on :action, 'message_action', 'unique_id' do |_action|
+            { text: 'message_action: exact match' }
           end
-        end
-      end
 
-      it 'performs action' do
-        post '/api/slack/action', payload: payload.merge(callback_id: 'action_id').to_json
-        expect(last_response.status).to eq 201
-        response = JSON.parse(last_response.body)
-        expect(response).to eq('text' => 'Success!')
-      end
-
-      it 'ignores an unhandled action' do
-        post '/api/slack/action', payload: payload.merge(callback_id: 'invalid').to_json
-        expect(last_response.status).to eq 204
-      end
-    end
-
-    context 'with both handlers' do
-      before do
-        SlackRubyBotServer::Events.configure do |config|
-          config.on :action, 'action_id' do |_action|
-            { text: 'Success!' }
+          config.on :action, 'message_action' do |_action|
+            { text: 'message_action: type match' }
           end
 
           config.on :action do |action|
-            { text: "Invoked action #{action[:payload][:callback_id]}." }
+            { text: "#{action[:payload][:type]}/#{action[:payload][:callback_id]}" }
           end
         end
       end
 
-      it 'performs action' do
-        post '/api/slack/action', payload: payload.merge(callback_id: 'action_id').to_json
+      it 'performs specific action' do
+        post '/api/slack/action', payload: payload.merge(type: 'message_action', callback_id: 'unique_id').to_json
         expect(last_response.status).to eq 201
         response = JSON.parse(last_response.body)
-        expect(response).to eq('text' => 'Success!')
+        expect(response).to eq('text' => 'message_action: exact match')
       end
 
-      it 'ignores an unhandled action' do
-        post '/api/slack/action', payload: payload.merge(callback_id: 'invalid').to_json
+      it 'performs default action' do
+        post '/api/slack/action', payload: payload.merge(type: 'message_action', callback_id: 'updated').to_json
         expect(last_response.status).to eq 201
         response = JSON.parse(last_response.body)
-        expect(response).to eq('text' => 'Invoked action invalid.')
+        expect(response).to eq('text' => 'message_action: type match')
+      end
+
+      it 'performs any action' do
+        post '/api/slack/action', payload: payload.merge(type: 'global_shortcut', action_id: 'action_id').to_json
+        expect(last_response.status).to eq 201
+        response = JSON.parse(last_response.body)
+        expect(response).to eq('text' => 'global_shortcut/action_id')
       end
     end
   end