+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 #### 0.2.0 (Next)
 
+* Include action type in actions callbacks - [@dblock](https://github.com/dblock).
 * Your contribution here.
 
 #### 0.1.0 (07/19/2020)